Remove usage of process.exit() when displaying error messages to the user
This is now handled in the bin script.

We still rely on knowing when we're synchronous so we can throw vs.
callback. but it's now possible to unit test negative cases such as bad
arguments with coverage reporting.

Added negative test cases for getUserConfig() and the new command.
Fixes some cases where we exited with a 0 code when arguments were missing.

Closes #23