Set default hostname to 'localhost' for webpackServer
When a hostname is not provided, a user is shown a message
`The app is running at HTTP://localhost:3000/`. This, however,
is not always correct. If the development is done within the crostini
container, the default hostname will be `http://penguin.linux.test:3000/'.

Setting the default hostname to 'localhost' makes everything work out of the box.